rename to sink, move utils to test utils
